refactor(options): move layout components into layout/ subdirectory
